requirements.txt
requests==2.26.0
uvicorn==0.20.0
fastapi==0.63.0
fastapi-utils==0.2.1
mysql-connector-python==8.0.30
mysqlclient==2.0.3
python-dotenv
psycopg2>=2.7,<3.0
pytz==2022.7.1
